backend/requirements.txt
# Python Processing Service - University Learning Platform
# Handles: OCR, text extraction, AI metadata generation, quiz generation

fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-dotenv==1.0.1

# File parsing
PyPDF2==3.0.1
pdfplumber==0.11.4
python-docx==1.1.2
python-pptx==1.0.2

# OCR
pytesseract==0.3.13
Pillow==10.4.0

# AI / LLM (Google Gemini)
google-genai>=1.0.0

# Database (for direct access if needed)
psycopg2-binary==2.9.9
sqlalchemy==2.0.35

# HTTP
httpx==0.27.2
